+++ packages/jaeger-ui/src/components/TracePage/TraceTimelineViewer/utils.js
@@ -1,3 +1,5 @@
+// @flow
+
 // Copyright (c) 2017 Uber Technologies, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,6 +13,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+import type { Span } from '../../../types/trace';
 
 export type ViewedBoundsFunctionType = (number, number) => { start: number, end: number };
 /**
@@ -27,7 +31,13 @@ export type ViewedBoundsFunctionType = (number, number) => { start: number, end:
  *                            relative to the `min`, `max`.
  * @returns {(number, number) => Object} Created view bounds function
  */
-export function createViewedBoundsFunc({ min, max, viewStart, viewEnd }) {
+export function createViewedBoundsFunc(viewRange: {
+  min: number,
+  max: number,
+  viewStart: number,
+  viewEnd: number,
+}) {
+  const { min, max, viewStart, viewEnd } = viewRange;
   const duration = max - min;
   const viewMin = min + viewStart * duration;
   const viewMax = max - (1 - viewEnd) * duration;
@@ -39,7 +49,7 @@ export function createViewedBoundsFunc({ min, max, viewStart, viewEnd }) {
    * @param  {number} end       The end of the sub-range.
    * @return {Object}           The resultant range.
    */
-  return (start, end) => ({
+  return (start: number, end: number) => ({
     start: (start - viewMin) / viewWindow,
     end: (end - viewMin) / viewWindow,
   });
@@ -54,7 +64,7 @@ export function createViewedBoundsFunc({ min, max, viewStart, viewEnd }) {
  *                        items.
  * @return {boolean}      True if a match was found.
  */
-export function spanHasTag(key, value, span) {
+export function spanHasTag(key: string, value: any, span: Span) {
   if (!Array.isArray(span.tags) || !span.tags.length) {
     return false;
   }
@@ -66,7 +76,7 @@ export const isServerSpan = spanHasTag.bind(null, 'span.kind', 'server');
 
 const isErrorBool = spanHasTag.bind(null, 'error', true);
 const isErrorStr = spanHasTag.bind(null, 'error', 'true');
-export const isErrorSpan = span => isErrorBool(span) || isErrorStr(span);
+export const isErrorSpan = (span: Span) => isErrorBool(span) || isErrorStr(span);
 
 /**
  * Returns `true` if at least one of the descendants of the `parentSpanIndex`
@@ -79,7 +89,7 @@ export const isErrorSpan = span => isErrorBool(span) || isErrorStr(span);
  *                                         the parent span will be checked.
  * @return     {boolean}  Returns `true` if a descendant contains an error tag.
  */
-export function spanContainsErredSpan(spans, parentSpanIndex) {
+export function spanContainsErredSpan(spans: Span[], parentSpanIndex: number) {
   const { depth } = spans[parentSpanIndex];
   let i = parentSpanIndex + 1;
   for (; i < spans.length && spans[i].depth > depth; i++) {
@@ -93,7 +103,7 @@ export function spanContainsErredSpan(spans, parentSpanIndex) {
 /**
  * Expects the first span to be the parent span.
  */
-export function findServerChildSpan(spans) {
+export function findServerChildSpan(spans: Span[]) {
   if (spans.length <= 1 || !isClientSpan(spans[0])) {
     return false;
   }